backend/requirements.txt
# FastAPI 核心
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# 数据处理 & AI
httpx>=0.28.0
pydantic>=2.10.0
email-validator>=2.0.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0

# 数据库
supabase>=2.11.0

# 图片生成
pillow>=11.0.0
qrcode>=8.0

# 金融数据
akshare>=1.16.0
apscheduler>=3.10.0  # 定时任务

# 配置
pyyaml>=6.0.0
